requirements.txt
Django==1.5
django-widget-tweaks==1.3
MySQL-python
argparse==1.2.1
django-mptt
django-mptt-admin
six==1.3.0
pil==1.1.7
wsgiref==0.1.2
